refactor: 지역성 고려하여 폴더구조 변경
