--- src/store/reducers/root.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import productReducer from './product';
-
-const rootReducer = combineReducers({
-  product: productReducer,
-});
-
-export default rootReducer;
